divided article view into smaller widgets
