+++ lib/views/article_view/article_view_logic.dart
@@ -1,0 +1,1 @@
+class ArticleViewManager {}
