+++ pubspec.yaml
@@ -30,7 +30,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   url_launcher: ^6.0.6
   firebase_core: ^1.0.1
-  provider: ^5.0.0
+  # provider: ^6.0.0
   cached_network_image: ^3.1.0
   intl: ^0.17.0
   share_plus: ^2.1.4
@@ -47,6 +47,8 @@ dependencies:
   webview_flutter: ^2.1.1
   google_fonts: ^2.1.0
   firebase_dynamic_links: ^2.0.10
+  flutter_math_fork: ^0.5.0
+  get_it: ^7.2.0
 
 dev_dependencies:
   flutter_test:
